pyproject.toml
[tool.poetry]
name = "chat-with-your-data-solution-accelerator"
version = "1.1.0"
description = "Chat with your data solution accelerator"
authors = []
readme = "README.md"
packages = [
    { include = "code" }
]

[tool.poetry.dependencies]
python = "^3.10"
azure-functions = "1.21.0"
streamlit = "1.39.0"
python-dotenv = "1.0.1"
azure-ai-formrecognizer = "3.3.3"
azure-storage-blob = "12.20.0"
azure-identity = "1.19.0"
flask = {extras = ["async"], version = "^3.0.3"}
openai = "1.48.0"
langchain = "0.2.16"
langchain-community = "0.2.17"
langchain-openai = "0.1.25"
requests = "2.32.3"
tiktoken = "0.8.0"
azure-storage-queue = "12.12.0"
beautifulsoup4 = "4.12.3"
fake-useragent = "1.5.1"
chardet = "5.2.0"
azure-search-documents = "11.6.0b1"
azure-ai-contentsafety = "1.0.0"
python-docx = "1.1.2"
azure-keyvault-secrets = "4.9.0"
pandas = "2.2.3"
azure-monitor-opentelemetry = "^1.6.2"
opentelemetry-instrumentation-httpx = "^0.48b0"
pillow = "10.4.0"
azure-mgmt-cognitiveservices = "^13.5.0"
jsonschema = "^4.23.0"
semantic-kernel = {version = "1.3.0", python = "<3.13"}
azure-ai-ml = "^1.21.1"
azure-cosmos = "^4.7.0"
azure-ai-documentintelligence = "1.0.0"
pymupdf = "^1.25.0"
azure-cognitiveservices-speech = "^1.41.1"
streamlit-extras = "^0.5.0"
streamlit-js-eval = "^0.1.7"
scipy = "^1.14.1"
azure-ai-textanalytics = "^5.3.0"
pdfkit = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "6.0.0"
flake8 = "7.1.1"
black = "24.10.0"
pre-commit = "4.0.1"
pytest_httpserver = "1.1.0"
trustme = "1.2.0"
jupyter = "1.1.1"
pytest-asyncio = "^0.24.0"

[tool.poetry.group.prompt-flow]
optional = true

[tool.poetry.group.prompt-flow.dependencies]
promptflow = {extras = ["azure"], version = "1.15.0"}

[tool.coverage.run]
omit = [
    "**/tests/*",
    "**/dist/*",
    "**/infra/*",
]

[tool.coverage.report]
skip_empty = true
include_namespace_packages = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
